autoGen()

Automatically translate network input data into LIM declaration files describing weighted and unweighted networks. Capable of translating multiple networks at once (via Excel '.xlsx' inputs of multiple sheets), or a single network at a time (via comma separated value '.csv'inputs for a single network).

When using .xlsx input files describing multiple networks, autoGen() simultaneously translates multiple network models into their respective LIM declaration files. This is advantageous when constructing multiple network models, e.g., for a time series. Function autoGen() requires two Workbooks, namely the ‘Network Data Input’ workbook (Workbook 1) and the ‘Adjacency Matrix’ workbook (Workbook 2) as input. Within both workbooks, the user specifies a unique worksheet name for each network model, which should match identically between the two workbooks.

As Microsoft Excel is proprietary and may not function on all operating systems on which R is used, autoGen() also accepts non-proprietary comma separated values (CSV) input files. As CSV files are limited to a single sheet per file, as opposed to Excel workbooks that can contain multiple sheets, the translation of input data from CSV files to LIM declaration files with autoGen() is limited to a single network at a time. Despite this limitation, the translation functionality from CSV files allows autoLIMR to be used in R on multiple operating systems, aligning with the power of R’s open source nature, availability, and computer operating system interpretability.

This function takes some action. It also attempts to create a few files in your working directory called data.R, under folders autoLIMR Weighted Network LIMfiles, and autoLIMR Unweighted Network LIMfiles. If data.R files cannot be created a warning is raised.

autoGen(
  net_data_input = "demo",
  adj_mat_input = "demo",
  NLNode = NULL,
  respiration = TRUE,
  respiration_element = "CO2",
  primary_producer = NULL,
  author = NULL,
  date = NULL,
  weighted = TRUE,
  force = TRUE
)

Arguments

net_data_input

the user-defined network input data workbook (.xlsx) or comma separated value (.csv) filename and file extension containing biomass and inequalities. If the argument is missing, the function will stop and return an error prompting the user to declare the appropriate input file.

adj_mat_input

the user-defined input data workbook (.xlsx) or comma separated value (.csv) filename and file extension containing adjacency matrices of internal network flows. If the argument is missing, the function will stop and return an error prompting the user to declare the appropriate input file.

NLNode

Character vector defining one or multiple non-living compartments e.g., NLNode = "Detritus", NLNode = c("Detritus", "Detritus2").

respiration

Logical. Is respiration included in the network model? If TRUE, the default, named respiration flows are defined from all defined living compartments (in the case of ecological networks) to an external carbon dioxide 'CO2' sink (e.g., Invertebrate_R: Invertebrate -> CO2). Additionally, if primary producers are included in the network (e.g., ecological networks), Gross Primary Production (GPP) flows are defined for each primary-producing compartment as the flow of energy/material from the external carbon dioxide sink to the compartment (e.g., Plant_GPP: CO2 -> Plant). If FALSE, no respiration flows are defined for living compartments.

respiration_element

Character vector. If NULL, the default, the external respiration element is defined as 'CO2'. This argument is only valid if argument respiration = TRUE.

primary_producer

Character vector defining one or multiple primary producer compartments (in the case of ecological networks) e.g., primary_producer = "Plant", primary_producer = c("Plant", "Plant2").

author

Character vector. LIM declaration file author name. Defaults to system user.

date

Character vector. Date of file creation. If NULL, the default, the date defaults to the system date.

weighted

Logical. Return weighted LIM declaration files? If NULL, the default, both weighted and unweighted LIM declaration files are returned. Defaults to TRUE.

force

If set to TRUE, files of data.R, under folders weighted_limfiles, and unweighted_limfiles will be created on the user's working directory. If this function is used in an interactive session the user will be asked whether or not data.R files should be created. The default value is FALSE.

Value

If argument weighted = TRUE, the default, a list of character vectors describing a 1) weighted and 2) unweighted network LIM declaration files. If argument weighted = FALSE, a list containing a single LIM declaration file describing an unweighted network. Each LIM declaration file is the standard input file to solve a single, or multiple plausible network configurations using linear inverse modelling (LIM) and Markov Chain Monte Carlo (MCMC) techniques using the autoLIMR function multi_net(), or with external R packages LIM and 'limSolve.

Examples

if (FALSE) { # \dontrun{
autoGen(
  net_data_input = "your_network_data_workbook.csv",
  adj_mat_input = "your_adjacency_matrix_data_workbook.csv",
  NLNode = "Detritus",
  primary_producer = "Plant",
  respiration = TRUE,
  respiration_element = "CO2",
  weighted = TRUE,
  force = TRUE
)

autoGen(
  net_data_input = "your_network_data_workbook.xlsx",
  adj_mat_input = "your_adjacency_matrix_data_workbook.xlsx",
  NLNode = c("Detritus", "Detritus2"),
  primary_producer = c("Plant", "Plant2"),
  respiration = TRUE,
  respiration_element = "CO2",
  weighted = TRUE,
  force = TRUE
)
} # }