autocorr_diag()

Function to calculate the autocorrelation function at specified lags for autoLIMR Markov chains. Based on coda::autocorr.diag() function (Plummer et al., 2020).

autocorr_diag(x, flow = NULL, ...)

Arguments

x

Object of class "multi_net_output" generated from autoLIMR::multi_net() function.

flow

Character argument. By default, flow = NULL, and the convergence diagnostics are returned for each Markov Chain variable. If convergence diagnostics are required for one Markov chain variable only, it can be specified as a character string e.g., flow = "Plant_GPP". We suggest leaving the default flow = NULL.

...

Other arguments from coda::autocorr.diag(). Check defaults in the coda package.

Value

Data frame with each row corresponding to one Markov chain variable, and columns describing autocorrelation function at specified lags.

Examples

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
#> Initial solution (x0) calculated using default LSEI algorithm (Haskell 
#> and Hanson 1981).  Solving multiple plausible network values (this may 
#> take a while)...
#> Warning: No equalities - setting type = 2
#>    user  system elapsed 
#>   0.236   0.025   0.260 
#> Packing multiple plausible values into network objects...
#> Setting edge attributes for 4 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Multiple plausible values solved and packed into network objects.
#> 
autocorr_diag(x = x)
#>                    Lag 0        Lag 1         Lag 5       Lag 10        Lag 50
#> Plant_GPP              1  0.062491912 -0.0213539650 -0.020972478  0.0010938124
#> Plant_R                1  0.236840710  0.0263857823  0.034201020 -0.0309151708
#> Invert_R               1  0.469965031 -0.0308981285  0.019244275  0.0221125878
#> Vert_R                 1  0.467954977  0.0612456672  0.020392696 -0.0414400396
#> DetNLNode_IM           1  0.363961199  0.0189750822  0.004908163  0.0375160340
#> Plant_IM               1  0.021385881 -0.0116185455  0.010640881 -0.0138226513
#> DetNLNode_EX           1  0.309693336 -0.0255243991 -0.039112705  0.0587468276
#> Plant_EX               1 -0.014524523  0.0161452846 -0.088878794  0.0070515050
#> Invert_EX              1 -0.029262182  0.0003617887 -0.024130947 -0.0208627765
#> Vert_EX                1 -0.006301674  0.0250874039  0.017661667  0.0217475477
#> DetNLNode_Q_Invert     1  0.467631908 -0.0074685649  0.015169587 -0.0160062440
#> Invert_Q_Vert          1  0.621681446  0.1266373571  0.034635410 -0.0002276117
#> Invert_U_DetNLNode     1  0.535223617  0.0523113088  0.012137559 -0.0276127472
#> Plant_Q_Invert         1  0.442933842 -0.0233697568 -0.005090353 -0.0181309519
#> Plant_U_DetNLNode      1  0.238956738 -0.0224394165 -0.029563830 -0.0129638706
#> Vert_U_DetNLNode       1  0.598959474  0.1222905535  0.022309351  0.0174335682
autocorr_diag(x = x, flow = "Plant_GPP")
#>           Lag 0      Lag 1       Lag 5      Lag 10      Lag 50
#> Plant_GPP     1 0.06249191 -0.02135396 -0.02097248 0.001093812