centralx0()

Function to solve multiple plausible networks with Markov Chain Monte Carlo methods, with a 'central' starting solution calculated from LIM::Xranges.

If the 'central' solution is not valid, the starting solution defaults to the Least Squares with Equalities and Inequalities (LSEI), suitable for under determined LIM problems.

centralx0(full_limfile, iter = NULL, jmp = NULL, x0 = "central", ...)

Arguments

full_limfile

LIM Declaration file built from LIM_Read().

iter

An integer defining the total number of iterations (samples) of the Markov Chain. If NULL, the default, the number of iterations is set at 3000. In our experience, we recommend setting a larger number of iterations (> 10,000) to ensure adequate sampling.

jmp

An integer defining the jump size (loosely interpreted as the 'distance' between iterative samples in the solution space). If NULL, the default, the jump size is internally calculated. If a value is provided, the value should be within the ranges of the flow magnitudes.

x0

A single string defining the starting solution algorithm. If NULL, the default, the first solution of the Markov Chain is solved with Least Squares with Equalities and Inequalities (LSEI). If "central", the first solution of the Markov Chain is solved with the central solution as per LIM:Xranges.

...

Further LIM::Xsample arguments.

Value

A list of multiple plausible network flow values. The first solution is solved with the central solution.