heidel_diag()

Function to calculate Heidelberger and Welch’s convergence diagnostic for autoLIMR Markov chains. Based on coda::heidel.diag() function (Plummer et al., 2020).

heidel_diag(x, flow = NULL, eps = 0.1, pvalue = 0.05, ...)

Arguments

x

Object of class "multi_net_output" generated from autoLIMR::multi_net() function.

flow

Character argument. By default, flow = NULL, and the convergence diagnostics are returned for each Markov Chain variable. If convergence diagnostics are required for one Markov chain variable only, it can be specified as a character string e.g., flow = "Plant_GPP". We suggest leaving the default flow = NULL.

eps

Target value for ratio of halfwidth to sample mean

pvalue

significance level to use

...

Other arguments from coda::heidel.diag()

Value

Data frame with each row corresponding to one Markov chain variable, and columns describing Heidelberger-Welch diagnostics. Includes 1) if the variable passes the Stationarity test ("HW.Stationarity.Test"), 2) starting iteration ("HW.Start.Iteration"), 3) p-value ("HW.p.value"), 4) halfwidth test evaluation ("HW.Halfwidth.Test"), 5) halfwidth mean ("HW.Mean"), and 6) halfwidth value ("HW.Halfwidth.Value").

Examples

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
#> Initial solution (x0) calculated using default LSEI algorithm (Haskell 
#> and Hanson 1981).  Solving multiple plausible network values (this may 
#> take a while)...
#> Warning: No equalities - setting type = 2
#>    user  system elapsed 
#>   0.233   0.004   0.237 
#> Packing multiple plausible values into network objects...
#> Setting edge attributes for 4 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Multiple plausible values solved and packed into network objects.
#> 
heidel_diag(x = x)
#>                    HW.Stationarity.Test HW.Start.Iteration HW.p.value
#> Plant_GPP                        Passed                  1  0.7875444
#> Plant_R                          Passed                  1  0.2502796
#> Invert_R                         Passed                  1  0.9869133
#> Vert_R                           Passed                  1  0.5309428
#> DetNLNode_IM                     Passed                  1  0.6025129
#> Plant_IM                         Passed                  1  0.1747371
#> DetNLNode_EX                     Passed                  1  0.2707615
#> Plant_EX                         Passed                  1  0.9915555
#> Invert_EX                        Passed                  1  0.2136867
#> Vert_EX                          Passed                  1  0.4795633
#> DetNLNode_Q_Invert               Passed                  1  0.3374883
#> Invert_Q_Vert                    Passed                  1  0.8595596
#> Invert_U_DetNLNode               Passed                  1  0.5434319
#> Plant_Q_Invert                   Passed                  1  0.4179101
#> Plant_U_DetNLNode                Passed                  1  0.5812836
#> Vert_U_DetNLNode                 Passed                  1  0.5945792
#>                    HW.Halfwidth.Test   HW.Mean HW.Halfwidth.Value
#> Plant_GPP                     Passed 918.54866          4.6488669
#> Plant_R                       Passed  89.42208          7.5090661
#> Invert_R                      Passed 401.10209         31.5175769
#> Vert_R                        Passed 309.88512         23.4767644
#> DetNLNode_IM                  Passed 327.96326         18.3144266
#> Plant_IM                      Passed  27.42803          0.8526757
#> DetNLNode_EX                  Passed 401.66904         16.8221841
#> Plant_EX                      Passed  23.27569          0.8722809
#> Invert_EX                     Passed  37.81810          1.5046593
#> Vert_EX                       Passed  10.76784          0.3916474
#> DetNLNode_Q_Invert            Passed 904.24518         26.7992519
#> Invert_Q_Vert                 Passed 754.53737         54.5840353
#> Invert_U_DetNLNode            Passed 450.44819         38.4237189
#> Plant_Q_Invert                Passed 739.66057         14.5762691
#> Plant_U_DetNLNode             Passed  93.61836          7.0006897
#> Vert_U_DetNLNode              Failed 433.88440         44.3565446
heidel_diag(x = x, flow = "Plant_GPP")
#>           HW.Stationarity.Test HW.Start.Iteration HW.p.value HW.Halfwidth.Test
#> Plant_GPP               Passed                  1  0.7875444            Passed
#>            HW.Mean HW.Halfwidth.Value
#> Plant_GPP 918.5487           4.648867