net_data_resp_flows()

Constructs respiration flows for each living compartment, if respiration is included in the model, and the external sink is 'CO2'

net_data_resp_flows(
  x,
  respiration = respiration,
  respiration_element = respiration_element,
  primary_producer = primary_producer,
  NLNode = NLNode
)

Arguments

x

Tidy network data input matrix.

respiration

Logical. Is respiration included in the network model? If TRUE, the default, named respiration flows are defined from all defined living compartments (in the case of ecological networks) to an external carbon dioxide 'CO2' sink (e.g., Invertebrate_R: Invertebrate -> CO2). Additionally, if primary producers are included in the network (e.g., ecological networks), Gross Primary Production (GPP) flows are defined for each primary-producing compartment as the flow of energy/material from the external carbon dioxide sink to the compartment (e.g., Plant_GPP: CO2 -> Plant). If FALSE, no respiration flows are defined for living compartments.

respiration_element

Character vector. If NULL, the default, the external respiration element is defined as 'CO2'. This argument is only valid if argument respiration = TRUE.

primary_producer

Character vector defining one or multiple primary producer compartments (in the case of ecological networks) e.g., primary_producer = "Plant", primary_producer = c("Plant", "Plant2").

NLNode

Character vector defining one or multiple non-living compartments e.g., NLNode = "Detritus", NLNode = c("Detritus", "Detritus2").

Value

Vector containing the respiration flows and vanity headings