raftery_diag()

Function to calculate Raftery and Lewis’s convergence diagnostic for autoLIMR Markov chains. Based on coda::raftery.diag() function (Plummer et al., 2020).

raftery_diag(
  x,
  flow = NULL,
  q = 0.025,
  r = 0.005,
  s = 0.95,
  converge.eps = 0.001,
  ...
)

Arguments

x

Object of class "multi_net_output" generated from autoLIMR::multi_net() function.

flow

Character argument. By default, flow = NULL, and the convergence diagnostics are returned for each Markov Chain variable. If convergence diagnostics are required for one Markov chain variable only, it can be specified as a character string e.g., flow = "Plant_GPP". We suggest leaving the default flow = NULL.

q

the quantile to be estimated.

r

the desired margin of error of the estimate.

s

the probability of obtaining an estimate in the interval (q-r,q+r).

converge.eps

Precision required for estimate of time to convergence.

...

Other arguments from coda::raftery.diag(). Check defaults in the coda package.

Value

Data frame with each row corresponding to one Markov chain variable, and columns describing various Raftery Lewis convergence diagnostics: 1) M, 2) N, 3) Nmin, 4) I. See coda package for further details.

Details

If the sample size (i.e., iterations) are too small, the function returns an error. We suggest keeping the sample size > 4000.

Examples

# \donttest{
fpath <- system.file(
  "example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)

set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 4000,
  jmp = 0.0001,
  x0 = NULL
)
#> Initial solution (x0) calculated using default LSEI algorithm (Haskell 
#> and Hanson 1981).  Solving multiple plausible network values (this may 
#> take a while)...
#>    user  system elapsed 
#>   0.048   0.005   0.053 
#> Packing multiple plausible values into network objects...
#> Multiple plausible values solved and packed into network objects.
#> 
# raftery_diag(x = x)
raftery_diag(x = x, flow = "Plant_GPP")
#>           RL.Burn.in.Length RL.Required.Sample.Size RL.Min.Sample.Size
#> Plant_GPP                16                   17980               3746
#>           RL.Dependence.Factor RL.Margin.of.error.r RL.Probablity.s
#> Plant_GPP                  4.8                0.005            0.95
#>           RL.Quantile.q
#> Plant_GPP         0.025
# }