summary_diag(): basic summary of autoLIMR Markov chain outputs.

Function to calculate summary statistics for autoLIMR Markov chains. Based on coda::summary() function (Plummer et al., 2020).

summary_diag(x, flow = NULL, xranges = TRUE, ...)

Arguments

x

Object of class "multi_net_output" generated from autoLIMR::multi_net() function.

flow

Character argument. By default, flow = NULL, and the convergence diagnostics are returned for each Markov chain variable. If convergence diagnostics are required for one Markov chain variable only, it can be specified as a character string e.g., flow = "Plant_GPP". We suggest leaving the default flow = NULL.

xranges

Logical argument, default = TRUE. Includes the LIM::Xranges calculated flow ranges for each flow.

...

Other arguments

Value

Data frame with each row corresponding to one Markov chain variable, and columns describing various summary statistics.

Examples

# Summary of object "multi_net_output" generated from autoLIMR::multi_net().
# The example LIM Declaration files can be found in the package folder
# "example_limfiles". If using a custom LIM Declaration file, users can
# save the LIM Declaration file in the working directory and refer directly
# within the function e.g., file = "mylimfile.R".
# If the LIM Declaration file is not saved in the
# working directory, the function will accept a valid file path.

fpath <- system.file("example_limfiles",
  "Winter_Weighted_Network_LIMfile.R",
  package = "autoLIMR"
)
set.seed(1)
x <- multi_net(
  file = fpath,
  iter = 1000,
  jmp = NULL,
  x0 = NULL
)
#> Initial solution (x0) calculated using default LSEI algorithm (Haskell 
#> and Hanson 1981).  Solving multiple plausible network values (this may 
#> take a while)...
#> Warning: No equalities - setting type = 2
#>    user  system elapsed 
#>   0.229   0.008   0.237 
#> Packing multiple plausible values into network objects...
#> Setting edge attributes for 4 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Setting edge attributes for 6 positive flows
#> Multiple plausible values solved and packed into network objects.
#> 
summary_diag(x = x) # MCMC Summary for all flows
#> Warning: Missing 'full_limfile' element in input. Xranges could not be calculated, using NA values instead.
#>                         Mean         SD   Naive SE Time-series SE        2.5%
#> Plant_GPP          918.54866  70.420364  2.2268874      2.3718709 744.6882244
#> Plant_R             89.42208  87.428075  2.7647185      3.8311562   1.9227369
#> Invert_R           401.10209 305.195437  9.6511271     16.0803964  12.0113913
#> Vert_R             309.88512 227.920044  7.2074646     11.9779410   9.9492986
#> DetNLNode_IM       327.96326 201.678953  6.3776485      9.3440952  17.6807029
#> Plant_IM            27.42803  13.757130  0.4350386      0.4350386   2.1702741
#> DetNLNode_EX       401.66904 196.945046  6.2279492      8.5827470  23.0996558
#> Plant_EX            23.27569  14.073441  0.4450413      0.4450413   1.2616938
#> Invert_EX           37.81810  24.276278  0.7676833      0.7676833   2.2080037
#> Vert_EX             10.76784   6.318867  0.1998201      0.1998201   0.5463246
#> DetNLNode_Q_Invert 904.24518 260.283455  8.2308855     13.6730877 390.5337013
#> Invert_Q_Vert      754.53737 425.146043 13.4442983     27.8489976 119.8886508
#> Invert_U_DetNLNode 450.44819 340.927517 10.7810747     19.6039382  15.0268334
#> Plant_Q_Invert     739.66057 146.050487  4.6185219      7.4368720 417.6180506
#> Plant_U_DetNLNode   93.61836  91.730941  2.9007871      3.5717805   1.6228471
#> Vert_U_DetNLNode   433.88440 358.228030 11.3281650     22.6308901  15.2094792
#>                           25%       50%        75%      97.5% Xranges_min
#> Plant_GPP          881.612957 941.67511  975.13822  997.82681          NA
#> Plant_R             24.368887  61.29494  125.90473  328.69998          NA
#> Invert_R           152.458339 332.30123  588.62223 1114.07102          NA
#> Vert_R             122.225968 264.60151  445.04499  845.98679          NA
#> DetNLNode_IM       156.180608 308.84914  496.99609  682.25167          NA
#> Plant_IM            16.162959  28.30840   39.53604   48.64815          NA
#> DetNLNode_EX       248.210808 422.30585  568.24045  695.90688          NA
#> Plant_EX            11.149950  22.47995   35.06857   48.29837          NA
#> Invert_EX           16.498856  35.41266   59.76489   79.36183          NA
#> Vert_EX              5.295244  10.69225   16.15186   21.55852          NA
#> DetNLNode_Q_Invert 707.652698 903.19736 1111.19042 1348.40387          NA
#> Invert_Q_Vert      425.402289 695.84913 1020.04262 1715.17433          NA
#> Invert_U_DetNLNode 182.304945 372.32087  662.53938 1202.70806          NA
#> Plant_Q_Invert     649.459670 764.24466  853.95336  951.21523          NA
#> Plant_U_DetNLNode   28.449378  65.72605  129.87481  356.77284          NA
#> Vert_U_DetNLNode   144.681581 348.11186  631.88231 1292.33980          NA
#>                    Xranges_max Xranges_central nvar niter nchain thin
#> Plant_GPP                   NA              NA   16  1000      1    1
#> Plant_R                     NA              NA   16  1000      1    1
#> Invert_R                    NA              NA   16  1000      1    1
#> Vert_R                      NA              NA   16  1000      1    1
#> DetNLNode_IM                NA              NA   16  1000      1    1
#> Plant_IM                    NA              NA   16  1000      1    1
#> DetNLNode_EX                NA              NA   16  1000      1    1
#> Plant_EX                    NA              NA   16  1000      1    1
#> Invert_EX                   NA              NA   16  1000      1    1
#> Vert_EX                     NA              NA   16  1000      1    1
#> DetNLNode_Q_Invert          NA              NA   16  1000      1    1
#> Invert_Q_Vert               NA              NA   16  1000      1    1
#> Invert_U_DetNLNode          NA              NA   16  1000      1    1
#> Plant_Q_Invert              NA              NA   16  1000      1    1
#> Plant_U_DetNLNode           NA              NA   16  1000      1    1
#> Vert_U_DetNLNode            NA              NA   16  1000      1    1
summary_diag(x = x, flow = "Plant_GPP") # MCMC summary for flow "Plant_GPP"
#> Warning: Missing 'full_limfile' element in input. Xranges could not be calculated, using NA values instead.
#>               Mean       SD Naive SE Time-series SE     2.5%     25%      50%
#> Plant_GPP 918.5487 70.42036 2.226887       2.371871 744.6882 881.613 941.6751
#>                75%    97.5% Xranges_min Xranges_max Xranges_central nvar niter
#> Plant_GPP 975.1382 997.8268          NA          NA              NA   16  1000
#>           nchain thin
#> Plant_GPP      1    1